python3.12/django/app_22/pyproject.toml
[tool.poetry]
name = "app-22"
version = "0.1.0"
description = ""
authors = ["noskill <golanghack@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
kafka-python = "^2.0.2"
channels = "^4.1.0"
channels-redis = "^4.2.0"
pytest-django = "^4.8.0"
drf-yasg = "^1.21.7"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
